agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mlops-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    # Web framework — async-first, lightweight, designed for microservice APIs
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",

    # AWS SDK
    "boto3>=1.34",
    "botocore>=1.34",

    # LangChain — agent orchestration and tool use
    "langchain>=0.2",
    "langchain-aws>=0.1",           # LangChain ↔ Bedrock bridge (ChatBedrock)
    "langchain-community>=0.2",

    # LlamaIndex — RAG pipeline and OpenSearch vector store integration
    "llama-index>=0.10",
    "llama-index-vector-stores-opensearch>=0.1",

    # OpenSearch client + AWS SigV4 auth for AOSS
    "opensearch-py>=2.5",
    "requests-aws4auth>=1.2",
    "requests>=2.31",

    # HTTP scraping
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",

    # Config — reads .env files and environment variables into a typed Settings class
    "pydantic>=2.7",
    "pydantic-settings>=2.2",

    # Structured logging — human-readable in dev, JSON in production
    "structlog>=24.0",

    # OpenTelemetry — distributed tracing (OTLP → ADOT sidecar → X-Ray on EKS)
    "opentelemetry-sdk>=1.24",
    "opentelemetry-exporter-otlp-proto-grpc>=1.24",
    "opentelemetry-instrumentation-fastapi>=0.45b0",
    "opentelemetry-instrumentation-httpx>=0.45b0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.14",
    "pytest-cov>=5.0",
    "moto[bedrock]>=5",     # mock boto3/Bedrock calls in unit tests
    "ruff>=0.4",            # fast linter + formatter
    "mypy>=1.10",           # static type checking
]

[tool.hatch.build.targets.wheel]
packages = ["src/agent"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]   # pycodestyle, pyflakes, isort, pyupgrade

[tool.mypy]
strict = true
disallow_any_explicit = true      # Fails the build if someone explicitly types `Any`
disallow_any_generics = true      # Fails if someone uses an untyped generic like `list` instead of `list[str]`
python_version = "3.11"
ignore_missing_imports = true     # some llama-index/langchain stubs are incomplete

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["src/agent"]
omit = ["*/tests/*"]
